add more searchtests + rework reversibleint: add implicit conversion and assignment operators

+++ oscar-algo/src/main/java/oscar/algo/reversible/ReversibleSetIndexedArray.java
@@ -76,7 +76,7 @@ public class ReversibleSetIndexedArray extends AbstractSetIndexedArray implement
 
 	@Override
 	public int getSize() {
-		return size.getValue();
+		return (int) size.getValue();
 	}
 
 	@Override
